requirements.txt
numpy
pandas
mysql-connector-python
pymysql
python-dotenv
scikit-learn
#-e .